+++ backend/scripts/dev_op/utils.ts
@@ -20,3 +20,24 @@ export async function promptBool(question: string): Promise<boolean> {
     rl.close();
     return ans === 'yes' || ans === 'y' || ans === '';
 }
+
+const projectRoot = `${__dirname}/../../`
+export const AZTEC_DEPLOYED_FOLDER_PATH = `${projectRoot}/scripts/deploy/aztec/aztecDeployments/`
+export const EVM_DEPLOYMENT_FOLDER_PATH = `${projectRoot}/ignition/deployments`
+
+
+export function getAztecDeployedAddressesFolderPath(chainId: bigint) {
+    return `${AZTEC_DEPLOYED_FOLDER_PATH}/${Number(chainId)}`
+}
+
+export function getAztecDeployedAddressesFilePath(chainId: bigint) {
+    return `${getAztecDeployedAddressesFolderPath(chainId)}/deployed_addresses.json`
+}
+
+export function getEvmDeployedAddressesFolderPath(chainId: bigint) {
+    return `${EVM_DEPLOYMENT_FOLDER_PATH}/chain-${Number(chainId)}`
+}
+
+export function getEvmDeployedAddressesFilePath(chainId: bigint) {
+    return `${getEvmDeployedAddressesFolderPath(chainId)}/deployed_addresses.json`
+}
